fuzz: cleanup utility_fuzz_test and add header string filter
