+++ test/common/http/common_fuzz.cc
@@ -1,0 +1,26 @@
+#include "common_fuzz.h"
+
+#include "common/common/macros.h"
+
+namespace Envoy {
+
+std::string filterInvalidCharacters(absl::string_view string) {
+  std::string filtered;
+  filtered.reserve(string.length());
+  for (const char& c : string) {
+    switch (c) {
+    case '\0':
+      FALLTHRU;
+    case '\r':
+      FALLTHRU;
+    case '\n':
+      filtered.push_back(' ');
+      break;
+    default:
+      filtered.push_back(c);
+    }
+  }
+  return filtered;
+}
+
+} // namespace Envoy
