+++ test/common/http/common_fuzz.h~
@@ -1,0 +1,7 @@
+#include <functional>
+
+namespace Envoy {
+
+
+
+}
