+++ test/common/http/utility_fuzz.proto
@@ -18,7 +18,6 @@ message UtilityTestCase {
   oneof utility_selector {
     string parse_query_string = 1;
     CookiesKey parse_cookie_value = 2;
-    CookiesKey has_set_cookie = 3;
     GetLastAddressFromXff get_last_address_from_xff = 4;
     string extract_host_path_from_uri = 5;
   }
